# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Manufacturing management API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemType } from "./activityItemType";

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  title: string;
  subtitle: string;
  amount?: number | null;
  timestamp: Date;
}