# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Manufacturing management API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryAttendanceToday } from "./dashboardSummaryAttendanceToday";
import type { DashboardSummaryCopperTrend } from "./dashboardSummaryCopperTrend";
import type { DashboardSummaryMonthlyPaymentsItem } from "./dashboardSummaryMonthlyPaymentsItem";

export interface DashboardSummary {
  totalEmployees: number;
  activeEmployees: number;
  totalSalaryPaid: number;
  totalSalaryPending: number;
  totalAdvancesGiven: number;
  copperCurrentPrice?: number | null;
  copperAveragePrice?: number | null;
  copperTrend: DashboardSummaryCopperTrend;
  attendanceToday: DashboardSummaryAttendanceToday;
  monthlyPayments: DashboardSummaryMonthlyPaymentsItem[];
}