# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Manufacturing management API
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  name: string;
  position: string;
  phone?: string | null;
  dailyWage: number;
  monthlySalary: number;
  joiningDate: Date;
  active: boolean;
  createdAt: Date;
}