# lib/api-zod/src/generated/types/employeeBalance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Manufacturing management API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeBalance {
  employeeId: number;
  employeeName: string;
  position: string;
  monthlySalary: number;
  totalEarned: number;
  totalPaid: number;
  totalAdvances: number;
  remainingDue: number;
  daysPresent: number;
  daysAbsent: number;
}