# lib/api-zod/src/generated/types/salaryPayment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Manufacturing management API
 * OpenAPI spec version: 0.1.0
 */
import type { SalaryPaymentStatus } from "./salaryPaymentStatus";

export interface SalaryPayment {
  id: number;
  employeeId: number;
  employeeName: string;
  amount: number;
  paymentDate: Date;
  periodMonth: string;
  status: SalaryPaymentStatus;
  notes?: string | null;
  createdAt: Date;
}